requirements.txt
pyannote.audio
torch
mediapipe
transformers
vosk
opencv-python
decord
gdown